rename parted -> device
